Redirect to myEYPD
added a login redirect for all users. This can be changed at some point to filter login redirects based on user role. The eypd_get_my_bookings_url() function is used in the nav links but doesn't seem to resolve to the correct URL in this context redirecting to /events. I am assuming that is due to a load order issue. Not a huge fan of hard coding /events on the end of this url.